Work more like the upstream
Modifying the dial() function to use the
upstream's idea of the DialerContext when it's available,
and using what I think is a reasonable default when it's not.

It may not be a reasonable default.

In addition, make the getConn() function either use an existing
DialContext, or a default TLS DialContext. If the first connection
attempt fails with a TLS failure, then bomb out. Otherwise
try to make one more connection with the DialContext deleted, which
should then use default behavior from dial().

This sort of tests with `go test -v -skip TestUnixSocket -skip
TestDiscoveryUnixSocket -skip TestFakeServer`, but that is skipping a
fair amount. It does break the TestFakeServer test case, since the
server will first receive TLS negotation, which it isn't prepared for